requirements.txt
pyttsx3
pyautogui
requests
SpeechRecognition
google-generativeai
python-dotenv
